QVSource example
Hi,

Thanks for the post.
I have a similar scenario. But wondering if you could provide an example on how to use the variable as in example the $(vRootUrl) but this need to have a response of another webservice response which I was able to get using QVSource connector.

Thanks heaps in advance.

Regards,
Sam.